simple-audio-video-transcription-on-aws/code/requirements.txt
awslambdaric==2.0.11
boto3==1.34.144
botocore==1.34.144
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
idna==3.7
jmespath==1.0.1
jwcrypto==1.5.6
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.3
s3transfer==0.10.1
simplejson==3.18.4
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.1
